refactor: add utils for apply coupon
